refactor(builder): centralize auto-join builder creation
Move shared auto-join builder bootstrapping into
AutoJoinQueryBuilderTrait and reuse it across integration paths.

This change:

- adds newAutoJoinBuilder() to the shared builder trait
- updates AutoJoinTrait to delegate newEloquentBuilder()
- updates QueryJoinerTrait to share the same builder factory path
- moves describeAutoJoinPath() to QueryJoinerTrait so model-defined
  path support is available on models using manual query integration

This removes duplicated builder setup and makes subquery compilation
rely on the same package builder contract in both tests and real apps.
